uodate

+++ VivaCityWebApi/VivaCityWebApi.Common/DAO/RessourceDao.cs
@@ -3,8 +3,8 @@
 	{
 		public int Id { get; set; } 
 		public RessourceItemDao RessourceItem { get; set; } = null!;
-		public double Nbr { get; set; } 
-		public double Max { get; set; }
+		public int Nbr { get; set; } 
+		public int Max { get; set; }
 		public int? VillageId { get; set; }
 		public int? RessourceItemId { get; set; }
 	}
