+++ VivaCityWebApi/VivaCityWebApi.Common/DTO/RessourceDto.cs
@@ -2,8 +2,8 @@
 	public class RessourceDto {
 		public int Id { get; set; } 
 		public RessourceItemDto RessourceItem { get; set; } = null!;
-		public double Nbr { get; set; } 
-		public double Max { get; set; }
+		public int Nbr { get; set; } 
+		public int Max { get; set; }
 		//public int? RessourceItemId { get; set; }
 		
 		public int? VillageId { get; set; }
